+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.15.0 (Unreleased)
 
+FEATURES
+
+* Added `sensu_secret` resource for managing Sensu secrets and data source for retrieving Sensu secrets
+
 ## 0.14.0 (July 22, 2024)
 
 IMPROVEMENTS